+++ src/app/static/src/app/store/projects/actions.ts
@@ -11,13 +11,14 @@ import {CurrentProject, Project, VersionDetails, VersionIds} from "../../types";
 
 export interface versionPayload {
     version: VersionIds,
-    name: string,
+    name?: string,
     note?: string
 }
 
 export interface projectPayload {
     projectId: number,
-    name: string
+    name?: string,
+    note?: string
 }
 
 export interface ProjectsActions {
@@ -33,6 +34,8 @@ export interface ProjectsActions {
     userExists: (store: ActionContext<ProjectsState, RootState>, email: string) => Promise<boolean>
     cloneProject: (store: ActionContext<ProjectsState, RootState>, payload: CloneProjectPayload) => void
     renameProject: (store: ActionContext<ProjectsState, RootState>, projectPayload: projectPayload) => void
+    updateVersionNote: (store: ActionContext<ProjectsState, RootState>, versionPayload: versionPayload) => void
+    updateProjectNote: (store: ActionContext<ProjectsState, RootState>, payload: projectPayload) => void
 }
 
 export interface CloneProjectPayload {
@@ -85,9 +88,13 @@ export const actions: ActionTree<ProjectsState, RootState> & ProjectsActions = {
             .postAndReturn<string>("/project/", qs.stringify({name}));
     },
 
-    async getProjects(context) {
+    async getProjects(context, isLoading = true) {
         const {commit} = context;
-        commit({type: ProjectsMutations.SetLoading, payload: true});
+
+        if (isLoading) {
+            commit({type: ProjectsMutations.SetLoading, payload: true});
+        }
+
         await api<ProjectsMutations, ProjectsMutations>(context)
             .withSuccess(ProjectsMutations.SetPreviousProjects)
             .withError(ProjectsMutations.ProjectError)
@@ -117,6 +124,20 @@ export const actions: ActionTree<ProjectsState, RootState> & ProjectsActions = {
             }
         }
     },
+    
+    async updateVersionNote(context, versionPayload: versionPayload) {
+        const {dispatch} = context
+        const {projectId, versionId} = versionPayload.version
+        const note = versionPayload.note
+
+        await api<ProjectsMutations, ErrorsMutation>(context)
+            .ignoreSuccess()
+            .withError(`errors/${ErrorsMutation.ErrorAdded}` as ErrorsMutation, true)
+            .postAndReturn(`/project/${projectId}/version/${versionId}/note`, qs.stringify({note}))
+            .then(() => {
+                dispatch("getProjects", false);
+            });
+    },
 
     async newVersion(context, note) {
         const {state} = context;
@@ -128,6 +149,19 @@ export const actions: ActionTree<ProjectsState, RootState> & ProjectsActions = {
             .withSuccess(ProjectsMutations.VersionCreated)
             .withError(`errors/${ErrorsMutation.ErrorAdded}` as ErrorsMutation, true)
             .postAndReturn(`project/${projectId}/version/?parent=${versionId}&note=${note}`)
+    },
+
+    async updateProjectNote(context, payload) {
+        const {dispatch} = context
+        const projectId = payload.projectId
+        const note = payload.note
+        await api<ProjectsMutations, ErrorsMutation>(context)
+            .ignoreSuccess()
+            .withError(`errors/${ErrorsMutation.ErrorAdded}` as ErrorsMutation, true)
+            .postAndReturn(`project/${projectId}/note`, qs.stringify({note}))
+            .then(() => {
+                dispatch("getProjects", false);
+            });
     },
 
     async loadVersion(context, version) {
@@ -193,12 +227,12 @@ export const actions: ActionTree<ProjectsState, RootState> & ProjectsActions = {
 
     async renameProject(context, projectPayload: projectPayload) {
         const {state, dispatch} = context;
-        const {projectId, name} = projectPayload
+        const {projectId, name, note} = projectPayload
 
         await api<ProjectsMutations, ErrorsMutation>(context)
             .ignoreSuccess()
             .withError(`errors/${ErrorsMutation.ErrorAdded}` as ErrorsMutation, true)
-            .postAndReturn(`/project/${projectId}/rename`, qs.stringify({name}))
+            .postAndReturn(`/project/${projectId}/rename`, qs.stringify({name: name, note: note}))
             .then(() => {
                 if (state.currentProject && state.currentProject.id === projectId) {
                     dispatch("getCurrentProject")
@@ -220,4 +254,3 @@ async function immediateUploadVersionState(context: ActionContext<ProjectsState,
             .postAndReturn(`/project/${projectId}/version/${versionId}/state/`, serialiseState(rootState));
     }
 }
-