+++ src/app/static/src/tests/integration/adr-dataset.itest.ts
@@ -7,7 +7,7 @@ import {actions as adrUploadActions} from "../../app/store/adrUpload/actions";
 import {SurveyAndProgramMutation} from "../../app/store/surveyAndProgram/mutations";
 import {getFormData} from "./helpers";
 import {ADRMutation} from "../../app/store/adr/mutations";
-import {UploadFile} from "../../app/types";
+import {UploadFile, Dataset} from "../../app/types";
 
 // this suite tests all endpoints that talk to the ADR
 // we put them in a suite of their own so that we can run
@@ -74,6 +74,21 @@ describe("ADR dataset-related actions", () => {
         expect(commit.mock.calls[4][0]).toBe(BaselineMutation.UpdateDatasetResources);
     });
 
+    it("can get releases", async () => {
+        const commit = jest.fn();
+
+        await adrActions.getDatasets({commit, rootState} as any);
+        const datasetId = commit.mock.calls[2][0]["payload"].find((dataset: Dataset) => dataset.title.includes("Antarctica")).id;
+
+        jest.resetAllMocks();
+        
+        await adrActions.getReleases({commit, rootState} as any, datasetId);
+        expect(commit.mock.calls[0][0].type).toBe(ADRMutation.SetReleases)
+        expect(commit.mock.calls[0][0].payload.length).toBeGreaterThan(0);
+        expect(commit.mock.calls[0][0].payload[0].name).toBeTruthy();
+        expect(commit.mock.calls[0][0].payload[0].id).toBeTruthy();
+    });
+
     it("can get userCanUpload when selected dataset organisation is set", async () => {
         const commit = jest.fn();
 
@@ -120,7 +135,6 @@ describe("ADR dataset-related actions", () => {
         expect(dataset).not.toBeNull()
 
         // 3. update dataset's organisation
-        commit.mockClear();
         const root = {
             ...rootState,
             baseline: {
@@ -129,13 +143,20 @@ describe("ADR dataset-related actions", () => {
                 }
             }
         };
-        const adr = {schemas, datasets: []}; //do not include datasets in adr state, the action will fetch them
-        await adrActions.getAndSetDatasets({commit, rootState, state: adr} as any, dataset.id)
 
         // 4. check can upload
         commit.mockClear();
-        const dispatch = jest.fn();
-        await adrActions.getUserCanUpload({commit, rootState: root, state: adr, dispatch} as any);
+        const dispatch = jest.fn().mockImplementation((type, payload) => {
+            if (type === "getDataset") {
+                root.baseline.selectedDataset = {
+                    id: payload.id,
+                    organization: {
+                        id: dataset.organization.id
+                    }
+                } as any
+            }
+        });
+        await adrActions.getUserCanUpload({commit, rootState: root, state: {schemas}, dispatch} as any);
         expect(commit.mock.calls[0][0].type).toBe(ADRMutation.SetUserCanUpload);
         expect(commit.mock.calls[0][0].payload).toBe(true);
     });
@@ -236,7 +257,7 @@ describe("ADR dataset-related actions", () => {
                 }
             }
         };
-        
+
         const uploadFilesPayload = [
             {
                 resourceType: "type1",
@@ -254,9 +275,8 @@ describe("ADR dataset-related actions", () => {
         expect(commit.mock.calls[1][0]["payload"]).toBe(1);
         expect(commit.mock.calls[2][0]["type"]).toBe("SetADRUploadError");
         expect(commit.mock.calls[2][0]["payload"]["error"]).toBe("OTHER_ERROR");
-        expect(dispatch.mock.calls.length).toBe(2);
-        expect(dispatch.mock.calls[0][0]).toBe("adr/getAndSetDatasets");
-        expect(dispatch.mock.calls[1][0]).toBe("getUploadFiles");
+        expect(dispatch.mock.calls.length).toBe(1);
+        expect(dispatch.mock.calls[0][0]).toBe("getUploadFiles");
     }, 7000);
 
 });