+++ src/app/static/src/tests/mocks.ts
@@ -42,6 +42,7 @@ import {initialModelCalibrateState, ModelCalibrateState} from "../app/store/mode
 import { HintrVersionState, initialHintrVersionState } from "../app/store/hintrVersion/hintrVersion";
 import {ADRState, initialADRState} from "../app/store/adr/adr";
 import {ADRUploadState, initialADRUploadState} from "../app/store/adrUpload/adrUpload";
+import {GenericChartState, initialGenericChartState} from "../app/store/genericChart/genericChart";
 
 export const mockAxios = new MockAdapter(axios);
 
@@ -161,6 +162,13 @@ export const mockColourScales = (props?: Partial<ColourScalesState>) => {
 export const mockProjectsState = (props?: Partial<ProjectsState>) => {
     return {
         ...initialProjectsState(),
+        ...props
+    }
+};
+
+export const mockGenericChartState =  (props?: Partial<GenericChartState>): GenericChartState => {
+    return {
+        ...initialGenericChartState(),
         ...props
     }
 };