+++ src/app/static/src/tests/modelCalibrate/mutations.test.ts
@@ -54,6 +54,8 @@ describe("ModelCalibrate mutations", () => {
         expect(state.error).toBeNull();
         expect(state.complete).toBe(false);
         expect(state.status).toStrictEqual({});
+        expect(state.generatingCalibrationPlot).toBe(false);
+        expect(state.calibratePlotResult).toBe(null);
     });
 
     it("CalibrateStatusUpdate sets status, resets error", () => {
@@ -112,6 +114,13 @@ describe("ModelCalibrate mutations", () => {
         expect(spy).toHaveBeenCalledWith(99);
     });
 
+    it("sets error stops calibration plot being generated", () => {
+        const state = mockModelCalibrateState({ generatingCalibrationPlot: true });
+        const error = mockError("TEST ERROR");
+        mutations[ModelCalibrateMutation.SetError](state, {payload: error});
+        expect(state.generatingCalibrationPlot).toBe(false);
+    });
+
     it("sets options", () => {
         const state = mockModelCalibrateState();
         const options = { "testOption": "testValue" };
@@ -123,5 +132,21 @@ describe("ModelCalibrate mutations", () => {
         const state = mockModelCalibrateState();
         mutations[ModelCalibrateMutation.PollingForStatusStarted](state, {payload: 99});
         expect(state.statusPollId).toBe(99);
+    });
+
+    it("sets calibration plot started and resets error and previous plot", () => {
+        const state = mockModelCalibrateState({error: mockError("TEST ERROR"), calibratePlotResult: {}});
+        mutations[ModelCalibrateMutation.CalibrationPlotStarted](state);
+        expect(state.generatingCalibrationPlot).toBe(true);
+        expect(state.calibratePlotResult).toBe(null);
+        expect(state.error).toBe(null);
+    });
+
+    it("sets calibration plot data", () => {
+        const state = mockModelCalibrateState({generatingCalibrationPlot: true});
+        const payload = { data: "TEST DATA" };
+        mutations[ModelCalibrateMutation.SetPlotData](state, {payload});
+        expect(state.calibratePlotResult).toStrictEqual({ payload });
+        expect(state.generatingCalibrationPlot).toBe(false);
     });
 });