+++ NEWS.md
@@ -1,7 +1,38 @@
-
-# hintr 1.41.0
+# hint 1.47.0
 
 * Add Portuguese translations
+
+# hint 1.46.0
+
+* Combine input data tabs
+
+# hint 1.45.0
+
+* Import releases from ADR: fetch and display release data from ADR
+
+# hint 1.44.1
+
+* Import releases from ADR: fetch data from correct ADR release
+
+# hint 1.44.0
+
+* Modify Rename dialog modal and add notes text area
+
+# hint 1.43.1
+
+* Bug fix: Importing from adr after fitting model should prompt new version behaviour
+
+# hint 1.43.0
+
+* Add note icon(s) and dialogs to project history for project/version
+
+# hint 1.42.0
+
+* Make login email case insensitive again
+
+# hint 1.41.0
+
+* Add bar chart to model calibrate page
 
 # hint 1.40.0